Change settings and move sb cron to hourly
