+++ app/Database/seeds/settings.yml
@@ -207,15 +207,8 @@
   options: ''
   type: boolean
   description: 'Only allow briefs to be created for bidded flights'
-- key: simbrief.expire_days
-  name: 'Simbrief Expire Time'
-  group: simbrief
-  value: 5
-  options: ''
-  type: number
-  description: 'Days after how long to remove unused briefs'
 - key: simbrief.expire_hours
-  name: 'Simbrief Expire Hours'
+  name: 'Simbrief Expire Time'
   group: simbrief
   value: 6
   options: ''
@@ -269,7 +262,7 @@
   value: false
   options: ''
   type: boolean
-  description: 'When enabled an aircraft can only be used only for one active SimBrief OFP'
+  description: 'When enabled, an aircraft can only be used for one active SimBrief OFP and Flight/Pirep'
 - key: pireps.duplicate_check_time
   name: 'PIREP duplicate time check'
   group: pireps
