+++ app/Providers/CronServiceProvider.php
@@ -2,11 +2,11 @@
 
 namespace App\Providers;
 
+use App\Cron\Hourly\ClearExpiredSimbrief;
 use App\Cron\Hourly\DeletePireps;
 use App\Cron\Hourly\RemoveExpiredBids;
 use App\Cron\Hourly\RemoveExpiredLiveFlights;
 use App\Cron\Nightly\ApplyExpenses;
-use App\Cron\Nightly\ClearExpiredSimbrief;
 use App\Cron\Nightly\NewVersionCheck;
 use App\Cron\Nightly\PilotLeave;
 use App\Cron\Nightly\RecalculateBalances;
@@ -31,7 +31,6 @@ class CronServiceProvider extends ServiceProvider
             SetActiveFlights::class,
             RecalculateStats::class,
             NewVersionCheck::class,
-            ClearExpiredSimbrief::class,
         ],
 
         CronWeekly::class => [
@@ -45,6 +44,7 @@ class CronServiceProvider extends ServiceProvider
             DeletePireps::class,
             RemoveExpiredBids::class,
             RemoveExpiredLiveFlights::class,
+            ClearExpiredSimbrief::class,
         ],
     ];
 }
